Split out map methods into own Map class
No changes to the algorithm at this point, just moving from ScoutMap
